mithila-art-classification/flask-backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
tensorflow==2.13.0
numpy==1.24.0
pillow==9.5.0
requests==2.31.0
gunicorn